requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --annotation-style=line --constraint=all-requirements.txt --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1   # via -c all-requirements.txt, aiohttp
aiohttp==3.12.13          # via -c all-requirements.txt, realtime
aiosignal==1.3.2          # via -c all-requirements.txt, aiohttp
annotated-types==0.7.0    # via -c all-requirements.txt, pydantic
anyio==4.9.0              # via -c all-requirements.txt, httpx, openai
attrs==25.3.0             # via -c all-requirements.txt, aiohttp
black==25.1.0             # via -c all-requirements.txt, -r requirements.in
certifi==2025.6.15        # via -c all-requirements.txt, httpcore, httpx
click==8.2.1              # via -c all-requirements.txt, black
deprecation==2.1.0        # via -c all-requirements.txt, postgrest
distro==1.9.0             # via -c all-requirements.txt, openai
dotenv==0.9.9             # via -c all-requirements.txt, -r requirements.in
flake8==7.3.0             # via -c all-requirements.txt, -r requirements.in
frozenlist==1.7.0         # via -c all-requirements.txt, aiohttp, aiosignal
gotrue==2.12.0            # via -c all-requirements.txt, supabase
h11==0.16.0               # via -c all-requirements.txt, httpcore
h2==4.2.0                 # via -c all-requirements.txt, httpx
hpack==4.1.0              # via -c all-requirements.txt, h2
httpcore==1.0.9           # via -c all-requirements.txt, httpx
httpx[http2]==0.28.1      # via -c all-requirements.txt, gotrue, openai, postgrest, python-telegram-bot, storage3, supabase, supafunc
hyperframe==6.1.0         # via -c all-requirements.txt, h2
idna==3.10                # via -c all-requirements.txt, anyio, httpx, yarl
iniconfig==2.1.0          # via -c all-requirements.txt, pytest
isort==6.0.1              # via -c all-requirements.txt, -r requirements.in
jiter==0.10.0             # via -c all-requirements.txt, openai
mccabe==0.7.0             # via -c all-requirements.txt, flake8
multidict==6.5.0          # via -c all-requirements.txt, aiohttp, yarl
mypy-extensions==1.1.0    # via -c all-requirements.txt, black
numpy==2.3.1              # via scipy
openai==1.90.0            # via -c all-requirements.txt, -r requirements.in
packaging==25.0           # via -c all-requirements.txt, black, deprecation, pytest
pathspec==0.12.1          # via -c all-requirements.txt, black
platformdirs==4.3.8       # via -c all-requirements.txt, black
pluggy==1.6.0             # via -c all-requirements.txt, pytest
postgrest==1.0.2          # via -c all-requirements.txt, supabase
propcache==0.3.2          # via -c all-requirements.txt, aiohttp, yarl
pycodestyle==2.14.0       # via -c all-requirements.txt, flake8
pydantic==2.11.7          # via -c all-requirements.txt, gotrue, openai, postgrest
pydantic-core==2.33.2     # via -c all-requirements.txt, pydantic
pyflakes==3.4.0           # via -c all-requirements.txt, flake8
pygments==2.19.2          # via -c all-requirements.txt, pytest
pyjwt==2.10.1             # via -c all-requirements.txt, gotrue
pytest==8.4.1             # via -c all-requirements.txt, pytest-mock
pytest-mock==3.14.1       # via -c all-requirements.txt, gotrue
python-dateutil==2.9.0.post0  # via -c all-requirements.txt, realtime, storage3
python-dotenv==1.1.0      # via -c all-requirements.txt, dotenv
python-telegram-bot==22.1  # via -c all-requirements.txt, -r requirements.in
realtime==2.4.3           # via -c all-requirements.txt, supabase
scipy==1.16.0             # via -r requirements.in
six==1.17.0               # via -c all-requirements.txt, python-dateutil
sniffio==1.3.1            # via -c all-requirements.txt, anyio, openai
storage3==0.11.3          # via -c all-requirements.txt, supabase
strenum==0.4.15           # via -c all-requirements.txt, supafunc
supabase==2.15.3          # via -c all-requirements.txt, -r requirements.in
supafunc==0.9.4           # via -c all-requirements.txt, supabase
tqdm==4.67.1              # via -c all-requirements.txt, openai
typing-extensions==4.14.0  # via -c all-requirements.txt, anyio, openai, pydantic, pydantic-core, realtime, typing-inspection
typing-inspection==0.4.1  # via -c all-requirements.txt, pydantic
websockets==14.2          # via -c all-requirements.txt, realtime
yarl==1.20.1              # via -c all-requirements.txt, aiohttp
